HW6/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pytz==2021.3
urllib3==1.26.8
Werkzeug==2.0.3
requests==2.27.1
python-dateutil==2.8.2